[JENKINS-10255]: Make Mercurial changelog based on comparison with previous build
Previously, the plugin was saving the results of "hg incoming" using a template as the changelog.  This worked as expected in trivial cases (only one node, workspace never removed), but led to some unexpected behaviors in more complicated cases as called out in the ticket.  The new behavior should improve this behavior by instead basing the changelog on the output of a "hg log" between the current revision and the previously built revision.  This should work regardless of whether there are multiple nodes involved, the current workspace was/wasn't up-to-date, or whether it was a clone or an update.  A couple unit tests are included to reproduce the problem, and verify that it's fixed.