Add a check to the tests for "hg" actually being available
On some environments, it's possible to have hg configured properly in the path
for command-line use, but not in graphical applications (such as most IDEs).
Mac OS X is an example of one such environment.  In these cases, the test should
now fail with a more informative message, rather than with a generic
"No such file or directory" message that doesn't explain what the problem is.