Add an option to perform a full clone
By default, the plugin performs clones with the `--rev` option.
According to the Mercurial man page, this:

    To pull only a subset of changesets, specify one or more
    revisions identifiers with -r/--rev or branches with
    -b/--branch. The resulting clone will contain only the
    specified changesets and their ancestors.

On very large repositories, the ancestor calculation can put a large
load on the Mercurial server, resulting in slow clone times.  Doing a
full clone is actually much quicker, even though the clone is larger.

This adds an option for doing a full clone.  To stay backwards
compatible, full clones are off by default.