FIX repository cache w/o sharing fails on Windows
getRepoLocation() returns a path with a "/" separator, but on Windows hgrc contains "\" paths for local clones.
Using ini4j we can simply search for the "default" path in the file and replace it without checking its value. After a clone it has been set by Mercurial and must match what we gave it, short of the directory separator.